数据库整数统一映射为Integer

+++ src/main/java/com/vova/phoenix/model/dto/request/TaskConfigFilterReq.java
@@ -6,6 +6,6 @@ import lombok.Data;
 public class TaskConfigFilterReq {
     private String appPlatform;
 
-    private Byte taskType;
+    private Integer taskType;
 
 }
